+++ tests/testthat/test-create.R
@@ -4,11 +4,7 @@ library(ollamar)
 test_that("create function works with basic input", {
     skip_if_not(test_connection(logical = TRUE), "Ollama server not available")
 
-    expect_error(create("mario"))
-    expect_error(create("mario", modelfile = "abc", path = "abc"))
-    expect_error(create("mario", path = "abc"))
-
-    resp <- create("mario", "FROM llama3\nSYSTEM You are mario from Super Mario Bros.")
+    resp <- create("mario", "deepseek-r1:1.5b")
     expect_s3_class(resp, "httr2_response")
     expect_equal(resp$status_code, 200)
     expect_true(model_avail("mario"))