+++ tests/testthat/test-pull.R
@@ -31,7 +31,7 @@ test_that("pull function works", {
     expect_vector(result$body)
 
     # correct model
-    result <- pull('llama3', stream = TRUE)
+    result <- pull('snowflake-arctic-embed:22m', stream = TRUE)
     # for this endpoint, even when stream = FALSE, the response is chunked)
     expect_true(httr2::resp_headers(result)$`Transfer-Encoding` == "chunked")
     expect_s3_class(result, "httr2_response")
@@ -50,6 +50,8 @@ test_that("pull function works", {
     expect_s3_class(pull('sdafd', stream = FALSE, insecure = FALSE), "httr2_response")
     expect_s3_class(pull('sdafd', stream = TRUE, insecure = TRUE), "httr2_response")
     expect_s3_class(pull('sdafd', stream = TRUE, insecure = FALSE), "httr2_response")
+
+    delete("snowflake-arctic-embed:22m")
 
 })